Split out types so flow doesn't choke on its sandwiches
