+++ src/types/JsdocTypes.js
@@ -1,50 +1,3 @@
-/* @flow */
-/* eslint-disable no-unused-vars */
-
-import React, {PropTypes} from 'react';
-import {List} from 'immutable';
-
-export const StampyPropTypes = {
-
-    className: PropTypes.string,
-
-    htmlProps: PropTypes.object,
-
-    onChange: PropTypes.func,
-
-    onChangeMulti: PropTypes.func,
-
-    onClick: PropTypes.func,
-
-    spruceModifier: PropTypes.oneOfType([
-        PropTypes.string,
-        PropTypes.object
-    ]),
-
-    spruceName: PropTypes.string
-};
-
-type ClassName = string;
-
-type HockApplier = (ComposedComponent: ReactClass<any>) => ReactClass<any>;
-
-type HtmlProps = Object;
-
-type ListOrArray = List<*> | Array<any>;
-
-type SpruceModifier = string | Object;
-
-type OnChangeMeta = {
-    event: Object,
-    element: Object
-}
-
-type OnChange = (newValue: string|boolean, meta: OnChangeMeta) => void;
-
-type OnChangeMulti = (newValue: Array<string|boolean>, meta: OnChangeMeta) => void;
-
-type OnClick = (event: Object) => void;
-
 /**
  * HTML class names to be added to the rendered component.
  *
